Implement endpoint for file uploads
Created an endpoint which a client can use to upload files to a
user-space. The endpoint takes the file as form-data, along with
two other fields:
1. user_email: To identify which user-space to upload the file to
2. path_from_base: Path from the base directory of the file to be
uploaded on the clien't machine. This is used to replicate the same
directory structure on the server as on the client machine from a
given base directory.