Implement endpoint to fetch signature file
Implemented an endpoint which can be used to get the signature
file for a paticular file in a user space. This endpoint will be
useful in pushing changes/updating a file.
The client will first request the signature file using this endpoint.
The client will then compute a delta file using this signature file
and send it to server using another endpoint which is yet to be made.
The server can now patch the file using this delta file.