Implement endpoint to push file changes
Created an endpoint which lets a client update a file in a user-space.
The client provides the user's email which is used to detect the user space.
In addition to this, the client provides the file path, relative to the
base directory, for the file which is to be updated, along with a delta
file which represents the differences between the client's version of the
file and the server's version of the file.
The client can get the signature file from the
/api/v1/user/{user_email}/path/{file_path: path} route.