Implement endpoint to pull changes in a file
Created the "/api/v1/pull-change" endpoint which is to be used by
a client to update a file on the client's machine to the same state
as the file on the server.
The client needs to provide the user's email to identify the user-space.
The client needs to provide the path of the file(relative to the base
directory on the client machine) it wants to update.
Additionally, the client also needs to provide a signature file for the
file it wants to update.
The server then uses this signature file to create a delta file and
responds back to the client with this delta file. The client can then
perform a patch operation using this delta file.