Store reverse deltas for version control
Every time a request to patch a file is made, a reverse delta is
created and stored.
First, a temporary file is created on which the patch operation is
applied to produce a patched version of a file.
Then, a delta between the patched version and the current version of
the file stored on the server is created, with the patched version
as the basis file.
This delta stores instructions on how to go from the patched version of
the file to the current version of the file stored on the server. Hence,
this is a reverse delta.

The /api/v1/reverse_delta_ids/{file_path: path} endpoint can be used to
fetch a list of the reverse delta's avaiable for a given file.

The /api/v1/download/reverse_delta/{file_path: str}/{reverse_delta_id: str}
endpoint can be used to download a particular reverse delta, identified
by a file path and the reverse delta id.

A client application can first make a call to the endpoint which provides
the list of available reverse delta's for a given file.
The client application can then make a call to the download endpoint to
download a reverse delta with a particular id.
Once the client application has a set of reverse delta's,
it can repeatedly apply the patch operation on the client's machine, thereby
rolling back to a previous version of the file.