BUG FIX: Incorrect Reverse Delta fixed
The signature file previously was produced for the non-updated
version and the reverse delta was created for this signature
against the non-update file.
To create a reverse delta, the signature must be produced for the
updated file since the delta file must contain instructions on how
the updated file can be converted to the non-updated file.
Fixed this bug by generating a signature for the updated file, leading
to generation of a correct reverse delta.